requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bitarray==3.0.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
ckzg==2.0.1
constantly==23.10.4
cryptography==43.0.1
cytoolz==1.0.0
daphne==4.1.2
Django==5.1.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
frozenlist==1.5.0
hexbytes==1.2.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.1.0
multidict==6.1.0
parsimonious==0.10.0
postgres==4.0
propcache==0.2.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pyunormalize==16.0.0
redis==5.0.8
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
service-identity==24.1.0
setuptools==75.1.0
solc==0.0.0a0
sqlparse==0.5.1
toolz==1.0.0
Twisted==24.7.0
txaio==23.1.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.2
web3==7.5.0
websockets==13.1
yarl==1.17.1
zope.interface==7.0.3
